myRecipe url 수정
